Separate album downloading logic from user display. Allow downloading individual user albums.
