+++ sas/templates/sas/download_pictures.jinja
@@ -1,13 +1,30 @@
-{% macro download_button() %}
-  <div x-show="pictures.length > 0" x-cloak>
-    <button
-      :disabled="isDownloading"
-      class="btn btn-blue"
-      @click="downloadZip()"
-    >
-      <i class="fa fa-download"></i>
-      {% trans %}Download all pictures{% endtrans %}
-    </button>
-    <progress x-ref="progress" x-show="isDownloading"></progress>
+{# Helper macro to create a download button for a
+  record of albums with alpine
+
+  This needs to be used inside an alpine environment.
+  Downloaded pictures will be `pictures` from the 
+  parent data store.
+
+  Note:
+    This requires importing `bundled/sas/pictures-download-index.ts`
+
+  Parameters:
+    name (str): name displayed on the button, will be translated
+ #}
+{% macro download_button(name) %}
+  <div x-data="pictures_download">
+    <div x-show="pictures.length > 0" x-cloak>
+      <button
+        :disabled="isDownloading"
+        class="btn btn-blue"
+        @click="downloadZip()"
+      >
+        <i class="fa fa-download"></i>
+        {%- if name != "" -%}
+          {% trans %}{{ name }}{% endtrans %}
+        {%- endif -%}
+      </button>
+      <progress x-ref="progress" x-show="isDownloading"></progress>
+    </div>
   </div>
 {% endmacro %}
